Cleaned up the code so that it compiles with -Wall -Wextra -Werror under clang
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/FreeMat5@4474 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libFM/Convert.hpp
@@ -9,10 +9,11 @@ namespace FM
   template <class T, class S>
   struct OpRangeConvert
   {
+    // FIXME - does using double here break int64?
     inline static void rangeConvert(T& y, const S& x) {
-      if (x < std::numeric_limits<T>::min())
+      if (double(x) < double(std::numeric_limits<T>::min()))
 	y = std::numeric_limits<T>::min();
-      else if (x > std::numeric_limits<T>::max())
+      else if (double(x) > double(std::numeric_limits<T>::max()))
 	y = std::numeric_limits<T>::max();
       else
 	y = (T)(x);
